chore: refactor OnboardingCohortRegister => CoffeeChatRegister
